L15 refactor: implemented encryption endpoints controller

+++ Homework/java-web/src/main/java/org/geekhub/hw11/entity/HistoryEntry.java
@@ -12,7 +12,7 @@ public record HistoryEntry(
     @NonNull String message,
     @NonNull String encrypted,
     @NonNull EncodingAlgorithm algorithm,
-    @NonNull Integer userId,
+    @NonNull Long userId,
     @NonNull EncodingOperation operation
 ) {
 }
